Adding fields to User and seller services
