+++ paymore/src/main/java/com/revature/paymore/service/UserService.java
@@ -1,4 +1,21 @@
 package com.revature.paymore.service;
 
+import com.revature.paymore.repository.UserRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+
+@Service
 public class UserService {
+
+    @Autowired
+    UserRepository userRepository;
+
+    public UserService(UserRepository userRepository){
+        this.userRepository =userRepository;
+    }
+
+    private static final Logger logger = LoggerFactory.getLogger(UserService.class);
 }
