--- paymore/src/main/java/service/UserService.java
@@ -1,20 +1,0 @@
-package service;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import repository.UserRepository;
-
-@Service
-public class UserService {
-
-    @Autowired
-    UserRepository userRepository;
-
-    public UserService(UserRepository userRepository){
-        this.userRepository =userRepository;
-    }
-
-    private static final Logger logger = LoggerFactory.getLogger(UserService.class);
-}
